requirements.txt
Flask
flask_cors
numpy
Keras
gevent
pillow
h5py
tensorflow
imutils
